chore(types): migrate from pyright to basedpyright (standard mode)
basedpyright is the workspace-canonical type checker per the
2026-05-17 strict-linting decisions. It is stricter than vanilla
pyright (97.8% typing-spec conformance vs lower); bundles its own
Node runtime via nodejs-wheel; ships a 'recommended' mode that
catches unannotated functions vanilla pyright skips.

Starting at typeCheckingMode = 'standard' for now; Task 8 upgrades
to 'recommended' once the easy wins from ruff expansion + pre-commit
additions have landed.

No behavior change. Suppression comments added inline where
third-party stubs lag the runtime (cv2, etc.); see the inline noqa
comments for the rationale of each.

+++ pd_book_tools/geometry/image_ops.py
@@ -76,10 +76,10 @@ def _tight_bbox_from_thresh(thresh: ndarray):
 
 def _connected_content_bbox_from_image_thresh(
     thresh: ndarray,
-    x1: int,
-    y1: int,
-    x2: int,
-    y2: int,
+    x1: float,
+    y1: float,
+    x2: float,
+    y2: float,
 ):
     """Return bbox for connected components that intersect the original ROI.
